requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
black==22.8.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.8.0
frozenlist==1.3.1
idna==3.4
jieba3k==0.35.1
joblib==1.2.0
lxml==4.9.1
multidict==6.0.2
mypy==0.981
mypy-extensions==0.4.3
newspaper3k==0.2.8
nltk==3.7
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
tinysegmenter==0.3
tldextract==3.3.1
tomli==2.0.1
tqdm==4.64.1
types-requests==2.28.11
types-urllib3==1.26.25
typing_extensions==4.3.0
urllib3==1.26.12
yarl==1.8.1
